pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "sfx-hit-finder"
version = "0.1.0"
description = "PoC: Peakfinder8 vs Vision Transformer for serial crystallography hit finding"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "h5py>=3.8",
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "tqdm",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
